refr(change): [[..slug]] deleted, one common layout.jsx created route groups removed as they are no longer needed (todo profile page)
