--- src/App.js
@@ -1,7 +1,0 @@
-import "../src/app/styles/global.css";
-import Welcome from "./app/(auth)/welcome/page";
-function App() {
-  return <Welcome></Welcome>;
-}
-
-export default App;
