--- src/app/(unauth)/login/index.css
@@ -1,54 +1,0 @@
-.login-form {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  padding: 20px;
-  align-items: center;
-  justify-content: start;
-  width: 300px;
-  height: 300px;
-  background-color: white;
-  border-radius: 5px;
-}
-.login-page {
-  display: flex;
-  flex-direction: column;
-  color: white;
-  align-items: center;
-  justify-content: center;
-}
-.login-input {
-  width: 100%;
-  height: 15%;
-  padding: 10px;
-  border-radius: 5px;
-  border: 0.5px solid rgb(203, 202, 202);
-}
-.login-button {
-  width: 100%;
-  height: 15%;
-  border-radius: 5px;
-  background-color: lightblue;
-  color: white;
-  font-size: 20px;
-  font-weight: 1000;
-  border: 0.5px solid lightblue;
-}
-.forgotten-password {
-  color: rgb(98, 157, 157);
-}
-.horizontal-divider {
-  width: 100%;
-  background-color: rgb(215, 215, 215);
-  height: 1px;
-}
-.create-account-button {
-  width: 70%;
-  height: 15%;
-  border-radius: 5px;
-  background-color: rgb(65, 179, 65);
-  color: white;
-  font-size: 20px;
-  font-weight: 300;
-  border: 0.5px solid rgb(65, 179, 65);
-}
