+++ src/app/blog/page.jsx
@@ -1,7 +1,7 @@
-import BlogPost from "../../components/BlogPost/BlogPost.js";
-import * as Mushrooms from "../../components/data/Mushrooms.js";
-import Header from "../../components/header/Header.js";
-import Footer from "../../components/footer/Footer.js";
+import BlogPost from "../components/BlogPost/BlogPost.js";
+import * as Mushrooms from "../components/data/Mushrooms.js";
+import Header from "../components/header/Header.js";
+import Footer from "../components/footer/Footer.js";
 import "./index.css";
 const MushroomsList = [
   Mushrooms.AmanitaMuscaria,
@@ -14,7 +14,6 @@ const MushroomsList = [
 export default function Blog() {
   return (
     <div>
-      <Header />
       <div className="blog-container">
         {MushroomsList.map((mushroom) => {
           return (
@@ -27,7 +26,6 @@ export default function Blog() {
           );
         })}
       </div>
-      <Footer />
     </div>
   );
 }
