+++ src/app/cart/page.jsx
@@ -1,16 +1,14 @@
-import Header from "../../components/header/Header";
-import Footer from "../../components/footer/Footer";
+import Header from "../components/header/Header";
+import Footer from "../components/footer/Footer";
 
 import "./index.css";
 
 export default function Cart() {
   return (
     <>
-      <Header />
       <div className="cart">
         <h1>Cart page is currently in development...</h1>
       </div>
-      <Footer />
     </>
   );
 }
