+++ src/app/layout.js
@@ -1,0 +1,38 @@
+"use client";
+import { UserProvider } from "@auth0/nextjs-auth0/client";
+import "./styles/global.css";
+import Header from "./components/header/Header";
+import Footer from "./components/footer/Footer";
+import { useEffect } from "react";
+
+// export const metadata = {
+//   title: "Next.js",
+//   description: "Generated by Next.js",
+// };
+
+export default function RootLayout({ children }) {
+  useEffect(() => {
+    const theme = window.matchMedia("(prefers-color-scheme: dark)").matches;
+    if (!localStorage.getItem("theme")) {
+      localStorage.setItem("theme", theme ? "dark" : "light");
+      localStorage.setItem("system", true);
+    }
+    document.documentElement.classList.toggle(
+      "dark",
+      localStorage.theme === "dark" ||
+        (!("theme" in localStorage) &&
+          window.matchMedia("(prefers-color-scheme: dark)").matches)
+    );
+  }, []);
+  return (
+    <html lang="en" id="root">
+      <UserProvider>
+        <body className="bg-neutral-200 dark:bg-neutral-900">
+          <Header></Header>
+          {children}
+          <Footer></Footer>
+        </body>
+      </UserProvider>
+    </html>
+  );
+}
