+++ src/app/page.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+// import Header from "../../components/header/Header.js";
+// import Footer from "../../components/footer/Footer.js";
+// import Content from "../../components/Content.js";
+import "./index.css";
+
+function Welcome() {
+  return (
+    <div className="bg-neutral-200 dark:bg-neutral-900">
+      <div className="wrapper"></div>
+    </div>
+  );
+}
+
+export default Welcome;
