+++ src/app/posts/[id]/index.css
