+++ src/app/products/[id]/index.css
