+++ src/app/settings/page.jsx
@@ -1,5 +1,5 @@
-import Header from "../../components/header/Header";
-import Footer from "../../components/footer/Footer";
+import Header from "../components/header/Header";
+import Footer from "../components/footer/Footer";
 
 import "./index.css";
